+++ app/controllers/staff/planned_disbursements_controller.rb
@@ -5,7 +5,6 @@ class Staff::PlannedDisbursementsController < Staff::BaseController
     @activity = Activity.find(params["activity_id"])
     @planned_disbursement = PlannedDisbursement.new
     @planned_disbursement.parent_activity = @activity
-    pre_fill_providing_organisation
     pre_fill_financial_quarter_and_year
 
     authorize @planned_disbursement
@@ -55,21 +54,12 @@ class Staff::PlannedDisbursementsController < Staff::BaseController
     params.require(:planned_disbursement).permit(
       :currency,
       :value,
-      :providing_organisation_name,
-      :providing_organisation_type,
-      :providing_organisation_reference,
       :receiving_organisation_name,
       :receiving_organisation_type,
       :receiving_organisation_reference,
       :financial_quarter,
       :financial_year,
     )
-  end
-
-  private def pre_fill_providing_organisation
-    @planned_disbursement.providing_organisation_name = @activity.providing_organisation.name
-    @planned_disbursement.providing_organisation_type = @activity.providing_organisation.organisation_type
-    @planned_disbursement.providing_organisation_reference = @activity.providing_organisation.iati_reference
   end
 
   private def pre_fill_financial_quarter_and_year