+++ app/models/planned_disbursement.rb
@@ -2,7 +2,7 @@ class PlannedDisbursement < ApplicationRecord
   include PublicActivity::Common
   PLANNED_DISBURSEMENT_BUDGET_TYPES = {"1": "original", "2": "revised"}
 
-  strip_attributes only: [:providing_organisation_reference, :receiving_organisation_reference]
+  strip_attributes only: :receiving_organisation_reference
 
   belongs_to :parent_activity, class_name: "Activity"
   belongs_to :report, optional: true
@@ -17,7 +17,8 @@ class PlannedDisbursement < ApplicationRecord
     :receiving_organisation_name,
     :receiving_organisation_type,
     :financial_quarter,
-    :financial_year
+    :financial_year,
+    :providing_organisation_reference
   validates :value, inclusion: {in: 0.01..99_999_999_999.00}
 
   def unknown_receiving_organisation_type?