+++ CHANGELOG.md
@@ -357,6 +357,7 @@
   through the application interface.
 - Planned disbursements are always original when created
 - Planned disbursement currency is always GBP
+- Planned disbursement providing organisation is always BEIS
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-20...HEAD
 [release-20]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-19...release-20